requirements.txt
streamlit
pandas
plotly
requests
beautifulsoup4
openpyxl
xlrd
python-dotenv
stripe
folium
streamlit-folium